update dependencies, indexing device names for voice assistance, implement Scheduler model and timer utilities ( not working in cluster mode yet )

+++ src/models/client.js
@@ -19,7 +19,7 @@ module.exports = function(Client) {
   Client.disableRemoteMethodByName('replaceOrCreate');
   Client.disableRemoteMethodByName('createChangeStream');
 
-  Client.cacheIterator = function*(filter) {
+  Client.cacheIterator = async function*(filter) {
     let iterator;
     if (filter && filter.match) {
       iterator = Client.iterateKeys(filter);
@@ -41,17 +41,19 @@ module.exports = function(Client) {
   };
 
   /**
-   * Update clients stored in cache
-   * @method module:Client.updateCache
+   * Delete clients stored in cache
+   * @method module:Client.deleteAll
    * returns {array} clients - Cached clients keys
    */
-  Client.updateCache = async () => {
+  Client.deleteAll = async () => {
     try {
       const clients = [];
-      logger.publish(4, `${collectionName}`, 'updateCache:req', '');
+      logger.publish(4, `${collectionName}`, 'deleteAll:req', '');
       for await (const key of Client.cacheIterator()) {
-        clients.push(key);
-        await Client.delete(key);
+        if (key && key !== null) {
+          clients.push(key);
+          await Client.delete(key);
+        }
       }
       return clients;
     } catch (error) {
